Fix: a foolish bug in add_tuple.

+++ miniob-server/storage/common/condition_filter.cpp
@@ -160,9 +160,9 @@ bool DefaultConditionFilter::filter(const Record &rec) const
     } break;
     case DATES: {
       //TODO 参考INTS进行大小段转换
-      int left = *(int *)   left_value;
-      int right = *(int *)  right_value;
-      cmp_result = left - right;
+      int left = *(int*)  left_value;
+      int right = *(int*) right_value;
+      cmp_result = (int)(left - right);
     } break;
     default: {
     }
